backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-pptx>=1.0.2
openai>=1.55.0
google-genai>=1.0.0
python-dotenv>=1.0.0
